% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fiskdatabasen_connect}
\alias{fiskdatabasen_connect}
\title{Fiskdatabasen connection}
\usage{
fiskdatabasen_connect()
}
\value{
A connection to the database
}
\description{
Connection to fiskdatabasen
}
\examples{
con <- fiskdatabasen_connect()
}
