% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_columns}
\alias{get_columns}
\title{Get columns}
\usage{
get_columns(con, table)
}
\arguments{
\item{con}{table}
}
\value{
A vector containing the names of the columns of the specific table
}
\description{
Get the columns from a specific table
}
\examples{
cols <- get_columns(con, "Redskap")
}
