% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{query_table}
\alias{query_table}
\title{Query table}
\usage{
query_table(connection, table, columns = NA, condition = NA)
}
\arguments{
\item{connection, }{table, columns, condition}
}
\value{
A single string
}
\description{
Function that query the table and return the expected dataframe
}
\examples{
alta_data <- query_table(connection = con, table="ENKELTFISK", columns = c("Lregnr","EnkeltfiskID", "Feltaar", "Dato","KjonnID","Vill_oppdrettetID"), condition = "Lregnr = 200212")
}
