% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_columns}
\alias{return_columns}
\title{Get column string}
\usage{
return_columns(columns)
}
\description{
Gives a SQL compliant string for filtering the columns. If no columns are selected, return ALL columns
}
